changes after merging 4.0 version
